Refactor backend to module-style, add kotlin test suite
Now backend is split into independent modules so that we won't have to
load all parts of the application while testing. It also has the benefit
that we can start our modules asynchronously.

There is now a separate backend configuration `TestConfig` for running
tests. It makes sure that no mock data is loaded before tests etc.